requirements.txt
absl-py==2.3.0
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
array-record==0.4.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
box==0.1.5
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
Columnar==1.3.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.0
cycler==0.12.1
datasets==3.6.0
debugpy @ file:///D:/bld/debugpy_1680755658559/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dill==0.3.8
dm-tree==0.1.8
ensure==1.0.4
etils==1.5.2
evaluate==0.4.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing==2.2.0
fastapi==0.115.14
filelock==3.13.1
flatbuffers==1.12
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.6.1
funcsigs==1.0.2
gast==0.4.0
gensim==4.3.3
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.73.1
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1701831845989/work
jax==0.4.30
jaxlib==0.4.30
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.3
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
namex==0.1.0
narwhals==1.45.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.2.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
promise==2.3
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
propcache==0.3.2
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1681775298734/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyMuPDF==1.26.3
pyparsing==3.2.3
python-box==7.3.2
python-dateutil==2.8.2
python-docx==1.2.0
pytz==2025.2
pywin32==304
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1679317071559/work
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.46.2
streamlit==1.46.1
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.9.3
tensorflow-estimator==2.9.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.17.2
tensorflow-text==2.9.0
termcolor==3.1.0
tf-keras==2.15.0
threadpoolctl==3.6.0
timeago==1.0.14
tokenizers==0.21.2
toml==0.10.2
toolz==1.0.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchinfo==1.8.0
torchvision==0.22.1+cu128
tornado @ file:///D:/bld/tornado_1681817629351/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1748959427/work
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
xgboost==2.1.4
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
